news_category_controler

+++ application/controllers/news_category_controller.php
@@ -61,10 +61,10 @@ class News_category_controller extends CI_Controller {
         if ($id && !$news_category->get($id)) {
             redirect($back);
         }
-        
+        if($news_category->get($id)){
         $news_category->delete();
         redirect($back);
-    
+        }
     }
     
     function add() {
@@ -81,39 +81,30 @@ class News_category_controller extends CI_Controller {
         $act = $this->input->get_post('act');
     
         $news_category = new NewsCategory();
-        $categories = NewsCategory::getTree();
+        //$categories = NewsCategory::getTree();
         $arr_parent = array();
         
         if ($act == ACT_SUBMIT) {
             
             $lang = Language::getList();
-            
             while($lang->fetchNext()) {
                 
-                if ($this->input->post('name_'.$lang->id)) {
-                    $news_category->name .= '<'.$lang->id.'>'.utf8_escape_textarea($this->input->post('name_'.$lang->id)).'</'.$lang->id.'>';
+                if ($this->input->post('name_'.$lang->code)) {
+                    $news_category->name .= '<'.$lang->code.'>'.utf8_escape_textarea($this->input->post('name_'.$lang->code)).'</'.$lang->code.'>';
                 }
                 
-                if ($this->input->post('description_'.$lang->id)) {
-                    $news_category->description .= '<'.$lang->id.'>'.utf8_escape_textarea($this->input->post('description_'.$lang->id)).'</'.$lang->id.'>';
+                if ($this->input->post('description_'.$lang->code)) {
+                    $news_category->description .= '<'.$lang->code.'>'.utf8_escape_textarea($this->input->post('description_'.$lang->code)).'</'.$lang->code.'>';
                 }
                 
             }
+//            $news_category->name = utf8_escape_textarea($this->input->post('name'));
+//            $news_category->description = utf8_escape_textarea($this->input->post('description'));
+            $news_category->id_parent = $this->input->post('parent');
+            $news_category->keyword = utf8_escape_textarea($this->input->post('keyword'));
+            $news_category->link = utf8_escape_textarea($this->input->post('link'));           
             
-            $news_category->link = utf8_escape_textarea($this->input->post('link'));
-            $news_category->keywords = utf8_escape_textarea($this->input->post('keyword'));
-            $news_category->id_parent = $this->input->post('id_parent');
-            
-            if ($news_category->id_parent) {
-                if (is_array($news_category->id_parent)) {
-                    $arr_parent = $news_category->id_parent;
-                    $news_category->id_parent = implode(',', $arr_parent);
-                }
-                else {
-                    $arr_parent[] = $news_category->id_parent;
-                }
-            }
-            
+                        
             if ($news_category->validateInput()) {
                 $news_category->insert();           
                 
@@ -129,7 +120,7 @@ class News_category_controller extends CI_Controller {
         Menu::getMenuTree($array_menus, $filter);
         
         $this->data['news_category'] = $news_category;
-        $this->data['categories'] = $categories;
+        //$this->data['categories'] = $categories;
         $this->data['arr_parent'] = $arr_parent;
         $this->data['array_menus'] = $array_menus;
         $this->data['cfer'] = $cfer;
@@ -176,18 +167,17 @@ class News_category_controller extends CI_Controller {
             
             while($lang->fetchNext()) {
                 
-                if ($this->input->post('name_'.$lang->id)) {
-                    $news_category->name .= '<'.$lang->id.'>'.utf8_escape_textarea($this->input->post('name_'.$lang->id)).'</'.$lang->id.'>';
+                if ($this->input->post('name_'.$lang->code)) {
+                    $news_category->name .= '<'.$lang->code.'>'.utf8_escape_textarea($this->input->post('name_'.$lang->code)).'</'.$lang->code.'>';
                 }
                 
-                if ($this->input->post('description_'.$lang->id)) {
-                    $news_category->description .= '<'.$lang->id.'>'.utf8_escape_textarea($this->input->post('description_'.$lang->id)).'</'.$lang->id.'>';
+                if ($this->input->post('description_'.$lang->code)) {
+                    $news_category->description .= '<'.$lang->code.'>'.utf8_escape_textarea($this->input->post('description_'.$lang->code)).'</'.$lang->code.'>';
                 }
                 
             }
             
-            
-            $news_category->keywords = utf8_escape_textarea($this->input->post('keyword'));
+            $news_category->keyword = utf8_escape_textarea($this->input->post('keyword'));
             $news_category->link = utf8_escape_textarea($this->input->post('link'));
             
             if ($this->input->post('id_parent')) {
@@ -224,6 +214,20 @@ class News_category_controller extends CI_Controller {
         $this->data['section'] = $section;
         
         $this->load->view('main', $this->data);
+        
+    }
+    
+    function deleteImage($id_news_category) {
+        
+        $back = base_url('news_category');
+        User::checkAccessable($this->session->userdata('userID'), 'news_category/deleteImage');
+        
+        $news_category = new NewsCategory();
+        
+        if (!$id_news_category || !$news_category->get($id_news_category)) {
+            redirect($back);
+        }
+        redirect('news_category/edit/'.$news_category->id);
         
     }
